security-toolkit/tool-12-dns-analyzer/requirements.txt
flask==3.0.0
flask-cors==4.0.0
dnspython==2.4.2
requests==2.31.0
reportlab==4.0.7
